[Finish]注文機能完成

+++ app/controllers/customers/orders_controller.rb
@@ -10,7 +10,7 @@ class Customers::OrdersController < ApplicationController
     if params[:order][:select_address] == "0"
       @order.post_code = current_customer.postal_code
       @order.address = current_customer.address
-      @order.name = current_customer.first_name + current_customer.last_name
+      @order.name = current_customer.last_name + current_customer.first_name
       #binding.pry
     elsif params[:order][:select_address] == "1"
       @address = Address.find(params[:order][:address_id])
@@ -31,25 +31,23 @@ class Customers::OrdersController < ApplicationController
 
   def create
     cart_items = current_customer.cart_items.all
-
-    order = Order.new(order_params)
-    order.customer_id = current_customer.id
-    order.save
-    cart_items.each do |cart_item|
-      order_detail = OrderDetail.new
-      order_detail.item_id = cart_item.item_id
-      order_detail.customer_id = cart_item.customer_id
-      order_detail.price = cart_item.item.price
-      order_detail.quantity = cart_item.amount
-      order_detail.save
+    @order = current_customer.orders.new(order_params)
+    if @order.save
+      cart_items.each do |cart_item|
+        order_detail = OrderDetail.new
+        order_detail.item_id = cart_item.item_id
+        order_detail.order_id = @order.id
+        order_detail.customer_id = cart_item.customer_id
+        order_detail.price = cart_item.item.price
+        order_detail.quantity = cart_item.amount
+        order_detail.save
+      end
+      redirect_to complete_orders_path
+      cart_items.destroy_all
+    else
+      @order = Order.new(order_params)
+      render :new
     end
-    redirect_to complete_orders_path
-    cart_items.destroy_all
-
-
-
-
-
   end
 
   def index
@@ -57,10 +55,16 @@ class Customers::OrdersController < ApplicationController
   end
 
   def show
+    @order = current_customer.orders.find(params[:id])
+　　@order_detail = OrderDetails.find(params[:id])
   end
 
   private
   def order_params
     params.require(:order).permit(:payment_method, :post_code, :name, :address, :shipping_cost, :total_payment)
   end
+
+  # def order_detail_params
+  #   params.require(:order_detail).permit(:item_id,:price,:quantity,:making_status,:order_id)
+  # end
 end
